+++ src/app/auth-error/error-content.tsx
