+++ src/app/auth-error/page.tsx
@@ -1,5 +1,5 @@
 import {getServerAuthSession} from "~/server/auth";
-import AuthErrorComp from "~/app/auth-error/authComp";
+import AuthErrorComp from "~/app/auth-error/error-content";
 
 export default async function AuthError() {
     const session = await getServerAuthSession()
