+++ src/app/lembaga/anggota-kegiatan/_components/table/mahasiswa-card-table.tsx
