+++ src/app/lembaga/anggota-kegiatan/page.tsx
@@ -1,7 +1,7 @@
 // Server Auth
 import { getServerAuthSession } from "~/server/auth";
 // Components Import
-import AnggotaComp from "~/app/lembaga/anggota-kegiatan/_components/anggotaComp";
+import AnggotaComp from "~/app/lembaga/anggota-kegiatan/_components/anggota-content";
 // Api Import
 import { api } from "~/trpc/server";
