+++ src/app/pencarian/[params]/page.tsx
@@ -1,7 +1,7 @@
 "use server"
 
 import React from 'react';
-import PencarianPage from "~/app/_components/pencarian/search-comp";
+import PencarianPage from "~/app/_components/pencarian/pencarian-content";
 import {getServerAuthSession} from "~/server/auth";
 import {api} from "~/trpc/server";
